PARTNER EXTENSION 1+2 (OPTIONAL)
PARTNER EXTENSION 1 (OPTIONAL)
With a partner, write a JavaScript program to find the most frequently occuring item of an array.

PARTNER EXTENSION 2 (OPTIONAL)
Find the missing number in the following shuffled integer array of 1 to 200

Assume you get the array below as input and you need to print the missing number. Please propose two solutions.